+++ TextureReplacer/Reflections.cs
@@ -40,27 +40,33 @@ namespace TextureReplacer
     {
       // List of all created reflection scripts.
       private static readonly List<Script> Scripts = new List<Script>();
-      private static int currentScript;
 
-      private readonly RenderTexture envMap;
-      private readonly Transform transform;
-      private readonly Renderer[] meshes;
-      private readonly bool[] meshStates;
-      private readonly bool isEva;
-      private readonly int interval;
+      private readonly ReflectionProbe probe;
 
-      private int counter;
-      private int currentFace;
-      private bool isActive = true;
-
-      public Script(Component part, int updateInterval)
+      public Script(Component part)
       {
-        envMap = new RenderTexture(reflectionResolution, reflectionResolution, 24) {
-          hideFlags = HideFlags.HideAndDontSave, wrapMode = TextureWrapMode.Clamp, dimension = TextureDimension.Cube
+        reflectionResolution = 1024;
+
+        probe = part.gameObject.AddComponent<ReflectionProbe>();
+        probe.mode = ReflectionProbeMode.Realtime;
+        probe.refreshMode = ReflectionProbeRefreshMode.EveryFrame;
+        probe.resolution = reflectionResolution;
+        probe.timeSlicingMode = ReflectionProbeTimeSlicingMode.IndividualFaces;
+        probe.nearClipPlane = 0.4f;
+        probe.shadowDistance = 0;
+        probe.blendDistance = 0;
+        probe.hdr = false;
+        probe.center = new Vector3(0, 100, 0);
+        // probe.size = new Vector3(100, 100, 100);
+        probe.cullingMask = (1 << BuildingsLayer) | (1 << PartsLayer) | (1 << KerbalsLayer) |
+                            (1 << AtmosphereLayer) | (1 << BodiesLayer) | (1 << SunLayer) | (1 << GalaxyLayer);
+        probe.realtimeTexture = new RenderTexture(probe.resolution, probe.resolution, 24) {
+          dimension = TextureDimension.Cube,
+          wrapMode = TextureWrapMode.Clamp
         };
 
-        transform = part.transform;
-        isEva = part.GetComponent<KerbalEVA>() != null;
+        Transform transform = part.transform;
+        bool isEva = part.GetComponent<KerbalEVA>() != null;
 
         if (isEva) {
           SkinnedMeshRenderer visor = transform.GetComponentsInChildren<SkinnedMeshRenderer>(true)
@@ -70,20 +76,13 @@ namespace TextureReplacer
             Material material = visor.material;
 
             material.shader = Instance.visorShader;
-            material.SetTexture(Util.CubeProperty, envMap);
+            material.SetTexture(Util.CubeProperty, probe.realtimeTexture);
             material.SetColor(Util.ReflectColorProperty, visorReflectionColour);
+
+            visor.reflectionProbeUsage = ReflectionProbeUsage.Simple;
+            visor.probeAnchor = visor.transform;
           }
         }
-
-        meshes = transform.GetComponentsInChildren<Renderer>();
-        meshStates = new bool[meshes.Length];
-        interval = updateInterval;
-
-        counter = Util.Random.Next(updateInterval);
-        currentFace = Util.Random.Next(6);
-
-        EnsureCamera();
-        Update(true);
 
         Scripts.Add(this);
       }
@@ -91,7 +90,8 @@ namespace TextureReplacer
       public void Destroy()
       {
         Scripts.Remove(this);
-        Object.DestroyImmediate(envMap);
+        Object.Destroy(probe.realtimeTexture);
+        Object.Destroy(probe);
       }
 
       public bool Apply(Material material, Shader shader, Color reflectionColour)
@@ -100,7 +100,7 @@ namespace TextureReplacer
 
         if (reflectiveShader != null) {
           material.shader = reflectiveShader;
-          material.SetTexture(Util.CubeProperty, envMap);
+          material.SetTexture(Util.CubeProperty, probe.realtimeTexture);
           material.SetColor(Util.ReflectColorProperty, reflectionColour);
           return true;
         }
@@ -110,123 +110,48 @@ namespace TextureReplacer
 
       public void SetActive(bool value)
       {
-        if (!isActive && value) {
-          Update(true);
-        }
-
-        isActive = value;
-      }
-
-      public static void UpdateScripts()
-      {
-        if (Scripts.Count == 0 || Time.frameCount % reflectionInterval != 0) {
-          return;
-        }
-
-        currentScript %= Scripts.Count;
-
-        int startScript = currentScript;
-        do {
-          Script script = Scripts[currentScript];
-          currentScript = (currentScript + 1) % Scripts.Count;
-
-          if (script.isActive) {
-            script.counter = (script.counter + 1) % script.interval;
-
-            if (script.counter == 0) {
-              script.Update(false);
-              break;
-            }
-          }
-        } while (currentScript != startScript);
-      }
-
-      private void Update(bool force)
-      {
-        int faceMask = force ? 0x3f : 1 << currentFace;
-
-        // Hide all meshes of the current part.
-        for (int i = 0; i < meshes.Length; ++i) {
-          meshStates[i] = meshes[i].enabled;
-          meshes[i].enabled = false;
-        }
-
-        Transform cameraTransform = camera.transform;
-
-        // Skybox.
-        cameraTransform.position = GalaxyCubeControl.Instance.transform.position;
-        camera.farClipPlane = 100.0f;
-        camera.cullingMask = 1 << 18;
-        camera.RenderToCubemap(envMap, faceMask);
-
-        // Scaled space.
-        cameraTransform.position = ScaledSpace.Instance.transform.position;
-        camera.farClipPlane = 3.0e7f;
-        camera.cullingMask = (1 << 10) | (1 << 23);
-        camera.RenderToCubemap(envMap, faceMask);
-
-        // Scene.
-        Vector3 pos = transform.position;
-        cameraTransform.position = isEva ? pos + 0.4f * transform.up : pos;
-        camera.farClipPlane = 60000.0f;
-        camera.cullingMask = (1 << 0) | (1 << 1) | (1 << 5) | (1 << 15) | (1 << 17);
-        camera.RenderToCubemap(envMap, faceMask);
-
-        // Restore mesh visibility.
-        for (int i = 0; i < meshes.Length; ++i) {
-          meshes[i].enabled = meshStates[i];
-        }
-
-        currentFace = (currentFace + 1) % 6;
+        probe.enabled = value;
       }
     }
 
     // Instance.
     public static Reflections Instance { get; private set; }
 
-    // Visor reflection feature.
-    public bool IsVisorReflectionEnabled { get; private set; }
+    // Reflection type.
+    public Type ReflectionType { get; set; }
     // Print names of meshes and their shaders in parts with TRReflection module.
     public bool LogReflectiveMeshes { get; private set; }
 
+    // World space.
+    private const int PartsLayer = 0;
+    private const int KerbalsLayer = 17;
+    private const int BuildingsLayer = 15;
+    // Scaled space.
+    private const int AtmosphereLayer = 9;
+    private const int BodiesLayer = 10;
+    private const int SunLayer = 23;
+    // Galaxy cube space.
+    private const int GalaxyLayer = 18;
+
     private static readonly Log log = new Log(nameof(Reflections));
+
+    // Real reflection resolution.
+    private static int reflectionResolution = 256;
+    private static Type globalReflectionType = Type.None;
+    // Reflection colour.
+    private static Color visorReflectionColour = Color.white;
+
     // Reflective shader map.
     private static readonly string[,] ShaderNameMap = {
-      {"KSP/Diffuse", "Reflective/Bumped Diffuse"}, {"KSP/Specular", "Reflective/Bumped Diffuse"},
-      {"KSP/Bumped", "Reflective/Bumped Diffuse"}, {"KSP/Bumped Specular", "Reflective/Bumped Diffuse"},
+      {"KSP/Diffuse", "Reflective/Bumped Diffuse"},
+      {"KSP/Specular", "Reflective/Bumped Diffuse"},
+      {"KSP/Bumped", "Reflective/Bumped Diffuse"},
+      {"KSP/Bumped Specular", "Reflective/Bumped Diffuse"},
       {"KSP/Alpha/Translucent", "Reflective/Bumped Diffuse"},
       {"KSP/Alpha/Translucent Specular", "Reflective/Bumped Diffuse"}
     };
 
-    // Render layers:
-    //  0 - parts
-    //  1 - RCS jets
-    //  5 - engine exhaust
-    //  9 - sky/atmosphere
-    // 10 - scaled space bodies
-    // 15 - buildings, terrain
-    // 17 - kerbals
-    // 18 - skybox
-    // 23 - sun
-    private static readonly float[] CullDistances = {
-      500.0f, 5.0f, 0.0f, 0.0f, 0.0f, 50.0f, 0.0f, 0.0f, //
-      0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f,    //
-      0.0f, 50.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f,   //
-      0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f     //
-    };
-
     private readonly Dictionary<Shader, Shader> shaderMap = new Dictionary<Shader, Shader>();
-    // Reflection camera.
-    private static Camera camera;
-    // Reflection type.
-    public Type ReflectionType { get; set; }
-    // Real reflection resolution.
-    private static int reflectionResolution = 128;
-    // Interval in frames for updating environment map faces.
-    private static int reflectionInterval = 1;
-    private static Type globalReflectionType = Type.None;
-    // Reflection colour.
-    private static Color visorReflectionColour = Color.white;
 
     // Reflective visor shader.
     private Shader visorShader;
@@ -242,17 +167,13 @@ namespace TextureReplacer
     /// </summary>
     public void ReadConfig(ConfigNode rootNode)
     {
-      bool isVisorReflectionEnabled = IsVisorReflectionEnabled;
       bool logReflectiveMeshes = LogReflectiveMeshes;
 
       Util.Parse(rootNode.GetValue("reflectionType"), ref globalReflectionType);
       Util.Parse(rootNode.GetValue("reflectionResolution"), ref reflectionResolution);
-      Util.Parse(rootNode.GetValue("reflectionInterval"), ref reflectionInterval);
-      Util.Parse(rootNode.GetValue("isVisorReflectionEnabled"), ref isVisorReflectionEnabled);
       Util.Parse(rootNode.GetValue("visorReflectionColour"), ref visorReflectionColour);
       Util.Parse(rootNode.GetValue("logReflectiveMeshes"), ref logReflectiveMeshes);
 
-      IsVisorReflectionEnabled = isVisorReflectionEnabled;
       LogReflectiveMeshes = logReflectiveMeshes;
     }
 
@@ -261,8 +182,6 @@ namespace TextureReplacer
     /// </summary>
     public void Load()
     {
-      IsVisorReflectionEnabled = false;
-
       string shadersFileName = Application.platform switch {
         RuntimePlatform.WindowsPlayer => "shaders.windows",
         RuntimePlatform.OSXPlayer     => "shaders.osx",
@@ -280,8 +199,6 @@ namespace TextureReplacer
           visorShader = shadersBundle.LoadAsset<Shader>("assets/visor.shader");
           if (visorShader == null) {
             log.Print("Visor shader missing in the asset file");
-          } else {
-            IsVisorReflectionEnabled = true;
           }
         }
       } catch (System.Exception ex) {
@@ -320,27 +237,9 @@ namespace TextureReplacer
 
     private void Destroy()
     {
-      if (camera != null) {
-        Object.DestroyImmediate(camera.gameObject);
-      }
-
       if (visorShader != null) {
         Object.DestroyImmediate(visorShader);
       }
-    }
-
-    private static void EnsureCamera()
-    {
-      if (camera != null) {
-        return;
-      }
-
-      camera = new GameObject("TRReflectionCamera", typeof(Camera)).GetComponent<Camera>();
-      camera.enabled = false;
-      camera.clearFlags = CameraClearFlags.Depth;
-      // Any smaller number and visors will reflect internals of helmets.
-      camera.nearClipPlane = 0.125f;
-      camera.layerCullDistances = CullDistances;
     }
 
     /**