+++ TextureReplacer/TREvaModule.cs
@@ -43,9 +43,8 @@ namespace TextureReplacer
         Personaliser.Instance.PersonaliseEva(part, kerbal, useEvaSuit);
       }
 
-      if (Reflections.Instance.ReflectionType == Reflections.Type.Real &&
-          Reflections.Instance.IsVisorReflectionEnabled) {
-        reflectionScript = new Reflections.Script(part, 1);
+      if (Reflections.Instance.ReflectionType == Reflections.Type.Real) {
+        reflectionScript = new Reflections.Script(part);
         reflectionScript.SetActive(useEvaSuit);
       }
     }